Abstract producing regions from the definition of areas
